refactor: React migration: Move chat welcome message into a content folder
