+++ frontend/src/content/aiAssistantWelcomeMessage.ts
